+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v2.1.0
+
+- Upgraded TypeScript to `4.6.4`.
+- Improved typing of `Object.property.hasOwnProperty` and `Object.hasOwn`. (https://github.com/uhyo/better-typescript-lib/issues/4)
+- Fixed typing of `JSON.stringify` so that its return type includes `undefined` when possible. (https://github.com/uhyo/better-typescript-lib/issues/6)
+- Tightened typing of Array higher-order functions. Predicates passed to `filter`, `every` and `some` now must return a value of `boolean` type. (https://github.com/uhyo/better-typescript-lib/issues/7)
+
 ## v2.0.0
 
 - Upgraded TypeScript to `4.5.2`.